backend/requirements.txt
django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.11
